refactor(hooks): Consolidate language and currency data fetching
