+++ js/src/hooks/useAvailableStoreCurrencies.js
@@ -1,30 +1,29 @@
 /**
- * External dependencies
- */
-import { useSelect } from '@wordpress/data';
-
-/**
  * Internal dependencies
  */
-import { STORE_KEY } from '~/data/constants';
+import useAvailableLanguagesCurrencies from './useAvailableLanguagesCurrencies';
 
-const selectorName = 'getAvailableLanguagesCurrencies';
+/**
+ * @typedef {import('~/data/selectors').MCCurrency} MCCurrency
+ */
 
+/**
+ * @typedef {Object} AvailableStoreCurrencies
+ * @property {Array<MCCurrency>|null} currencies Available currencies from the multilingual integration, or null before data is fetched.
+ * @property {boolean} hasFinishedResolution Whether the shared resolver has completed its request.
+ */
+
+/**
+ * Returns available store currencies from the multilingual integration (e.g. WCML).
+ * Delegates to useAvailableLanguagesCurrencies so both values share one resolver call.
+ *
+ * @return {AvailableStoreCurrencies} Available currencies and resolution status.
+ */
 const useAvailableStoreCurrencies = () => {
-	return useSelect( ( select ) => {
-		const selector = select( STORE_KEY );
+	const { currencies, hasFinishedResolution } =
+		useAvailableLanguagesCurrencies();
 
-		// Trigger the shared resolver that populates both languages and currencies.
-		selector[ selectorName ]();
-
-		return {
-			currencies: selector.getAvailableStoreCurrencies(),
-			hasFinishedResolution: selector.hasFinishedResolution(
-				selectorName,
-				[]
-			),
-		};
-	}, [] );
+	return { currencies, hasFinishedResolution };
 };
 
 export default useAvailableStoreCurrencies;
